+++ src/components/detail/Detail.tsx
@@ -1,0 +1,2 @@
+import { SimpleGrid, Box, Text, Divider } from "@chakra-ui/react";
+
