+++ src/components/schedule/Structure.tsx
@@ -1,0 +1,11 @@
+export interface Schedule {
+  id: number;
+  title: string;
+  date: number; // 仮候補日程を作成した日付
+  description: string;
+}
+
+export interface ScheduleListProps {
+  width?: string | number;
+  schedules: Schedule[];
+}
